Added a small demo of enlive and hiccup.
Added a smaller demo of enlive by itself, without any actual web
framework code around it.  Also added a demo of hiccup, which has some
features that I thought were in enlive, but are not.  Both are really
cool approaches to generating html using clojure code, one is based on
existing html files which are modified at runtime, the other is based
on a literal sexpr syntax for building html directly.